WIP commit transactions
